+++ src/modules/image/image.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class ImageService {
+
+    saveFile(file: Express.Multer.File): string {
+        return file.filename;
+    }
+}